requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.22
apscheduler>=3.10.4
requests>=2.32.0
beautifulsoup4>=4.12.3
groq>=0.11.0
cloudinary>=1.41.0
jinja2>=3.1.4
python-multipart>=0.0.9
pillow>=10.4.0
langgraph>=0.2.60
huggingface_hub>=1.10.1
python-dotenv>=1.0.1

